Update prettier config and move cli functions (#304)
* Update prettier and refactor


* Update require
