+++ tests/api/filter.test.js
@@ -1,4 +1,4 @@
-// Copyright (c) 2020 Red Hat, Inc.
+// Copyright Contributors to the Open Cluster Management project
 
 jest.retryTimes(global.retry, { logErrorsBeforeRetry: true })
